12/06/25
Improved explosion behaviour to change how it interacts with the player compared to other objects. Additionally, the enemies now 'die' when exploded. Next update greater changes to the explosion on player collision will be made to improve game feel and allow synergy with the dash already in the players kit