15/06/2025 - END OF SPRINT 2
The player can now throw  explosives that freeze in place when they touch the ground and that propel the player forwards when they collide with one. The distance that the explosive travels is variable and is affected by how the player is moving. Changes were made to remedy this, but the explosive still travels differently depending on player movement which is not ideal. This may be fixed in sprint 4.