15/06 - part 2 (start of Sprint 3)
Created this instance of the enemy class that will be used in the game once the AIController is complete. A class based playerHealth script was also coded which freezes the player when hp reaches 0 and functions exactly as intended. Next lesson the ai will be added followed by an item that gives the player health back and then UI that displays current health to the player.