17/06
Finished the practical component of sprint 3. The classes previously written this script were overhauled because I realised that using nested classes added no functional value to the assessment. Instead, the AI and playerhealth classes are now separate from the monobehaviours they were attached to. The monobehaviors are now responsible for creating an instance of the class and passing in values from the editor. The AI class functions mostly as intended, with a slight bug in the behaviour related to movement as it gets substantially slower after attacking the player. The next update will be the start of sprint 4 and will add assets to the game unless I decide to revisit the functionality of the UI.