18/06 - END OF SPRINT 3
I intended to just do theory today and move on to sprint 4, but I realised that it would be really cool to add projectile enemies to the game as well, so I created children classes of the AI class (melee and ranged) to create more enemy variety. This meant a new script had to be created which holds the enemy projectile and that changes had to be made to be made to the logic of how attacks are reset in the AI class and the way that the explosive interacts with enemies. additionally. imported some assets that will used next sprint